4-Implementation of comment entity:
I have created all classes for the Comment entity. It is working. I have implemented a 'post comment' function on the chat page for now, this will be changed today. I have implemented a request for 'comments' to the API but have not managed to render that info on the screen yet. though it logs to console.